Revert "[autoscaler][observability] Autoscaler emit cluster/pending logical resource metrics (#29820)"
This reverts commit 3eaf3beb03cb330122bf8ebce8efa320780da5ce.
